+++ packages/mobx-state-tree/src/types/utility-types/maybe.ts
@@ -10,7 +10,6 @@ import {
     ExtractC,
     ExtractS,
     ExtractT,
-    OptionalProperty,
     IStateTreeNode,
     RedefineIStateTreeNode
 } from "../../internal"
@@ -21,15 +20,10 @@ const optionalNullType = optional(nullType, null)
 /** @hidden */
 export interface IMaybeIType<IT extends IAnyType, C, O>
     extends IType<
-            ExtractC<IT> | C,
-            ExtractS<IT> | O,
-            | RedefineIStateTreeNode<
-                  ExtractT<IT>,
-                  IStateTreeNode<ExtractC<IT> | C, ExtractS<IT> | O>
-              >
-            | O
-        >,
-        OptionalProperty {}
+        ExtractC<IT> | C,
+        ExtractS<IT> | O,
+        RedefineIStateTreeNode<ExtractT<IT>, IStateTreeNode<ExtractC<IT> | C, ExtractS<IT> | O>> | O
+    > {}
 
 /** @hidden */
 export interface IMaybe<IT extends IAnyType> extends IMaybeIType<IT, undefined, undefined> {}
@@ -47,7 +41,7 @@ export interface IMaybeNull<IT extends IAnyType> extends IMaybeIType<IT, null | 
 export function maybe<IT extends IAnyType>(type: IT): IMaybe<IT> {
     if (process.env.NODE_ENV !== "production" && !isType(type))
         throw fail("expected a mobx-state-tree type as first argument, got " + type + " instead")
-    return union(type, optionalUndefinedType) as any
+    return union(type, optionalUndefinedType)
 }
 
 /**
@@ -60,5 +54,5 @@ export function maybe<IT extends IAnyType>(type: IT): IMaybe<IT> {
 export function maybeNull<IT extends IAnyType>(type: IT): IMaybeNull<IT> {
     if (process.env.NODE_ENV !== "production" && !isType(type))
         throw fail("expected a mobx-state-tree type as first argument, got " + type + " instead")
-    return union(type, optionalNullType) as any
+    return union(type, optionalNullType)
 }